[fix] merge conflict solve
Fix : merge conflict 해결하기 위해 파일 정리

TryAxios.ts파일 수정 및 Appinner파일 삭제
